# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION for Geotager",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "prosenjak.urban@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/locations/guess/{id}": {
            "post": {
                "tags": [
                    "Guesses"
                ],
                "summary": "Submit a guess for a location",
                "description": "Allows a user to guess the location's coordinates and calculates the error distance. Subtracts points based on the number of previous guesses.",
                "operationId": "guessLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location to guess",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "latitude": {
                                        "description": "Latitude of the guess",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "description": "Longitude of the guess",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guess submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/guesses/best": {
            "get": {
                "tags": [
                    "Guesses"
                ],
                "summary": "Retrieve the best guesses for the authenticated user",
                "description": "Returns a list of the user's guesses ordered by smallest error distance.",
                "operationId": "getBestGuesses",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of top guesses to retrieve",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved best guesses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guess"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/guesses/{id}": {
            "get": {
                "tags": [
                    "Guesses"
                ],
                "summary": "Retrieve the best guesses from each unique user for a specific location",
                "description": "Returns the best guess (smallest error distance) from each user for the specified location.",
                "operationId": "getBestGuessesByLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of top guesses to retrieve per user",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved best guesses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guess"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Location not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Guess": {
                "required": [
                    "id",
                    "latitude",
                    "longitude",
                    "error_distance",
                    "user_id",
                    "location_id"
                ],
                "properties": {
                    "id": {
                        "description": "Guess ID",
                        "type": "integer"
                    },
                    "latitude": {
                        "description": "Guess latitude",
                        "type": "number",
                        "format": "float",
                        "example": 37.77493
                    },
                    "longitude": {
                        "description": "Guess longitude",
                        "type": "number",
                        "format": "float",
                        "example": -122.41942
                    },
                    "error_distance": {
                        "description": "Error distance",
                        "type": "number",
                        "format": "float"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "location_id": {
                        "description": "Location ID",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Timestamp when guess was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when guess was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}